PE32: Replace manual PE checksum calculations with Win32 API call instead. PE64: Replace manual PE checksum calculations with Win32 API call instead.
I was trying to avoid using Win32 API calls in this project for the most part, however, this calculation has multiple conditions that can alter its result. Steamless was only producing proper checksums about 40% of the time as a result of this. Instead, the project will now use the system API call 'MapFileAndCheckSum' to calculate the checksum properly.